requirements.txt
discord.py
gTTS
python-dotenv
PyNaCl
